+++ src/option.c
@@ -168,4 +168,5 @@ void set_default_options(void)
    opt_set_int(OPT_DRIVER_VERBOSE, get_int_env("NVC_DRIVER_VERBOSE", 0));
    opt_set_int(OPT_JIT_INTRINSICS, get_int_env("NVC_JIT_INTRINSICS", 1));
    opt_set_int(OPT_VECTOR_INTRINSICS, get_int_env("NVC_VECTOR_INTRINSICS", 1));
+   opt_set_int(OPT_SHUFFLE_PROCS, 0);
 }