+++ NEWS.md
@@ -13,6 +13,8 @@
 - Implemented the "closely related record types" feature from VHDL-2019.
 - Implemented the "composition with protected types" feature from
   VHDL-2019.
+- The new `--shuffle` option runs processes in a random order which can
+  help to identify code that depends on a particular execution order.
 
 ## Version 1.10.2 - 2023-08-20
 - Fixed a crash due to an array bounds check being incorrectly optimised